There should not be a distinct header file.
