+++ opencog/guile/modules/ExecSCM.cc
@@ -5,7 +5,26 @@
  * Copyright (c) 2008, 2014, 2015 Linas Vepstas <linas@linas.org>
  */
 
-#ifdef HAVE_GUILE
+#include <opencog/guile/SchemeModule.h>
+
+namespace opencog {
+
+class ExecSCM : public ModuleWrap
+{
+	protected:
+		virtual void init(void);
+		static std::vector<FunctionWrap*>* _binders;
+	public:
+		ExecSCM(void);
+		~ExecSCM();
+};
+
+}
+
+extern "C" {
+void opencog_exec_init(void);
+};
+
 
 #include <cstddef>
 #include <opencog/atoms/base/Link.h>
@@ -14,8 +33,6 @@
 #include <opencog/atoms/execution/Instantiator.h>
 #include <opencog/atoms/reduct/FoldLink.h>
 #include <opencog/guile/SchemeModule.h>
-
-#include "ExecSCM.h"
 
 // ========================================================
 
@@ -90,4 +107,3 @@ void opencog_exec_init(void)
 	static ExecSCM exy;
 	exy.module_init();
 }
-#endif // HAVE_GUILE
